+++ featurebyte/api/base_table.py
@@ -69,6 +69,41 @@ class TableColumn(FeatureByteBaseModel, ParentMixin):
         return cast(ColumnInfo, column_info)
 
     @property
+    def cleaning_operations(self) -> List[CleaningOperation]:
+        """
+        Cleaning operations applied to the column of the table.
+
+        Returns
+        -------
+        List[CleaningOperation]
+            List of cleaning operations applied to the column of the table.
+
+        Examples
+        --------
+        Show the list of cleaning operations of the event table amount column after updating the critical
+        data info.
+
+        >>> event_table = catalog.get_table("GROCERYINVOICE")
+        >>> event_table["Amount"].update_critical_data_info(
+        ...    cleaning_operations=[
+        ...        fb.MissingValueImputation(imputed_value=0),
+        ...    ]
+        ... )
+        >>> event_table["Amount"].cleaning_operations
+        [MissingValueImputation(imputed_value=0, type=missing)]
+
+        Empty list of column cleaning operations of the event table amount column.
+
+        >>> event_table["Amount"].update_critical_data_info(cleaning_operations=[])
+        >>> event_table["Amount"].cleaning_operations
+        []
+        """
+        for column_cleaning_operations in self.parent.column_cleaning_operations:
+            if column_cleaning_operations.column_name == self.name:
+                return cast(List[CleaningOperation], column_cleaning_operations.cleaning_operations)
+        return []
+
+    @property
     def feature_store(self) -> FeatureStoreModel:
         """
         Feature store used by parent frame